Split what the parser rejects out of ExpressionParserTest
Adding the arithmetic operators and parentheses pushed the file from
769 lines to 1059, and about half of that was error cases: four
providers for invalid syntax, type errors, and where each of those
points, plus the tests reading them. They share nothing with the parse
and print cases but the class they sat in — the accepted expressions
are checked by round-trip, the rejected ones by the message and span
they produce.

They move to ExpressionParserErrorTest, leaving 471 lines here and 602
there. Each file now says in a docblock what it covers and points at
the other.

testSyntaxErrorLocation and testTypeErrorLocation were the same
twenty-six lines twice, differing in the exception they catch: the same
try/fail/catch, then the same four-field span comparison with the same
eight-argument message. The comparison becomes assertLocation() and
each test keeps only its own catch.
